backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
pillow==9.5.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
ultralytics==8.0.186
torch>=2.0.0
torchvision>=0.15.0
